requirements.txt
python-socketio==5.8.0
requests==2.31.0
